refactor(@jest-reporters): split utils.ts
