+++ packages/jest-reporters/package.json
@@ -33,11 +33,13 @@
     "jest-worker": "^28.0.2",
     "slash": "^3.0.0",
     "string-length": "^4.0.1",
+    "strip-ansi": "^6.0.0",
     "terminal-link": "^2.0.0",
     "v8-to-istanbul": "^9.0.0"
   },
   "devDependencies": {
     "@jest/test-utils": "^28.0.2",
+    "@tsd/typescript": "~4.6.2",
     "@types/exit": "^0.1.30",
     "@types/glob": "^7.1.1",
     "@types/graceful-fs": "^4.1.3",
@@ -49,7 +51,7 @@
     "@types/node-notifier": "^8.0.0",
     "jest-resolve": "^28.0.3",
     "mock-fs": "^5.1.2",
-    "strip-ansi": "^6.0.0"
+    "tsd-lite": "^0.5.1"
   },
   "peerDependencies": {
     "node-notifier": "^8.0.1 || ^9.0.0 || ^10.0.0"
