+++ packages/jest-reporters/src/index.ts
@@ -5,16 +5,14 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import formatTestPath from './formatTestPath';
 import getResultHeader from './getResultHeader';
 import getSnapshotStatus from './getSnapshotStatus';
 import getSnapshotSummary from './getSnapshotSummary';
-import {
-  formatTestPath,
-  getSummary,
-  printDisplayName,
-  relativePath,
-  trimAndFormatPath,
-} from './utils';
+import getSummary from './getSummary';
+import printDisplayName from './printDisplayName';
+import relativePath from './relativePath';
+import trimAndFormatPath from './trimAndFormatPath';
 
 export type {
   AggregatedResult,
@@ -28,10 +26,10 @@ export type {Config} from '@jest/types';
 export {default as BaseReporter} from './BaseReporter';
 export {default as CoverageReporter} from './CoverageReporter';
 export {default as DefaultReporter} from './DefaultReporter';
+export {default as GitHubActionsReporter} from './GitHubActionsReporter';
 export {default as NotifyReporter} from './NotifyReporter';
 export {default as SummaryReporter} from './SummaryReporter';
 export {default as VerboseReporter} from './VerboseReporter';
-export {default as GitHubActionsReporter} from './GitHubActionsReporter';
 export type {
   Reporter,
   ReporterOnStartOptions,
